Dep: Move manifest updating logic into separate class
