+++ spec/dependabot/file_updaters/go/dep_spec.rb
@@ -103,168 +103,13 @@ RSpec.describe Dependabot::FileUpdaters::Go::Dep do
           end
 
           it "includes the new requirement" do
+            expect(described_class::ManifestUpdater).
+              to receive(:new).
+              with(dependencies: [dependency], manifest: manifest).
+              and_call_original
+
             expect(updated_manifest_content).
               to include(%(version = ">= 1.0.0, < 4.0.0"))
-          end
-        end
-      end
-
-      context "when the requirement in the manifest has been deleted" do
-        let(:requirements) do
-          [{
-            file: "Gopkg.toml",
-            requirement: nil,
-            groups: [],
-            source: {
-              type: "default",
-              source: "github.com/dgrijalva/jwt-go"
-            }
-          }]
-        end
-
-        its(:length) { is_expected.to eq(1) }
-
-        describe "the updated manifest" do
-          subject(:updated_manifest_content) do
-            updated_files.find { |f| f.name == "Gopkg.toml" }.content
-          end
-
-          it "includes the new requirement" do
-            expect(updated_manifest_content).
-              to end_with(%(name = "github.com/dgrijalva/jwt-go"\n))
-          end
-        end
-      end
-
-      context "when a requirement is being added to the manifest" do
-        let(:manifest_fixture_name) { "no_version.toml" }
-        let(:lockfile_fixture_name) { "no_version.lock" }
-        let(:previous_requirements) do
-          [{
-            file: "Gopkg.toml",
-            requirement: nil,
-            groups: [],
-            source: {
-              type: "default",
-              source: "github.com/dgrijalva/jwt-go"
-            }
-          }]
-        end
-        let(:requirements) do
-          [{
-            file: "Gopkg.toml",
-            requirement: ">= 1.0.0, < 4.0.0",
-            groups: [],
-            source: {
-              type: "default",
-              source: "github.com/dgrijalva/jwt-go"
-            }
-          }]
-        end
-
-        its(:length) { is_expected.to eq(1) }
-
-        describe "the updated manifest" do
-          subject(:updated_manifest_content) do
-            updated_files.find { |f| f.name == "Gopkg.toml" }.content
-          end
-
-          it "includes the new requirement" do
-            expect(updated_manifest_content).
-              to end_with("  name = \"github.com/dgrijalva/jwt-go\"\n"\
-                          "  version = \">= 1.0.0, < 4.0.0\"\n")
-          end
-        end
-      end
-
-      context "when the tag in the manifest has changed" do
-        let(:manifest_fixture_name) { "tag_as_revision.toml" }
-        let(:lockfile_fixture_name) { "tag_as_revision.lock" }
-        let(:dependency_name) { "golang.org/x/text" }
-        let(:dependency_version) { "v0.3.0" }
-        let(:dependency_previous_version) { "v0.2.0" }
-        let(:requirements) do
-          [{
-            requirement: nil,
-            file: "Gopkg.toml",
-            groups: [],
-            source: {
-              type: "git",
-              url: "https://github.com/golang/text",
-              branch: nil,
-              ref: "v0.3.0"
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "Gopkg.toml",
-            groups: [],
-            source: {
-              type: "git",
-              url: "https://github.com/golang/text",
-              branch: nil,
-              ref: "v0.2.0"
-            }
-          }]
-        end
-
-        its(:length) { is_expected.to eq(1) }
-
-        describe "the updated manifest" do
-          subject(:updated_manifest_content) do
-            updated_files.find { |f| f.name == "Gopkg.toml" }.content
-          end
-
-          it "includes the new tag" do
-            expect(updated_manifest_content).to include(%(revision = "v0.3.0"))
-          end
-        end
-      end
-
-      context "when switching from a git revision to a release" do
-        let(:manifest_fixture_name) { "tag_as_revision.toml" }
-        let(:lockfile_fixture_name) { "tag_as_revision.lock" }
-        let(:dependency_name) { "golang.org/x/text" }
-        let(:dependency_version) { "0.3.0" }
-        let(:dependency_previous_version) { "v0.2.0" }
-        let(:requirements) do
-          [{
-            requirement: "^0.3.0",
-            file: "Gopkg.toml",
-            groups: [],
-            source: {
-              type: "default",
-              source: "golang.org/x/text"
-            }
-          }]
-        end
-        let(:previous_requirements) do
-          [{
-            requirement: nil,
-            file: "Gopkg.toml",
-            groups: [],
-            source: {
-              type: "git",
-              url: "https://github.com/golang/text",
-              branch: nil,
-              ref: "v0.2.0"
-            }
-          }]
-        end
-
-        its(:length) { is_expected.to eq(1) }
-
-        describe "the updated manifest" do
-          subject(:updated_manifest_content) do
-            updated_files.find { |f| f.name == "Gopkg.toml" }.content
-          end
-
-          it "includes the new tag" do
-            expect(updated_manifest_content).
-              to end_with("  name = \"golang.org/x/text\"\n"\
-                          "  version = \"^0.3.0\"\n")
           end
         end
       end
